Delete mailbox, it isn't a mailbox
So after learning a bit more about UVM, a Queue is a raw collection
type, while Mailbox wraps it with functionality to ensure multiple
writers and readers are handled fairly and (thread) safely. cocotb's
Queue is already that.

I will do something else here, but I'm not sure yet. cocotb's Queue is
missing functionality I think it needs to have, but adding this
functionality to cocotb's Queue has some caveats.