Add Queue
This extends the functionality of cocotb.triggers.Queue and
asyncio.Queue on which it was based by adding support for peeking and
support for directly interacting with the read and write-side locks.

Acquiring the read and write locks allows the user to wait until a side
of the queue is available and they have exclusive access to it, allowing
them to anywhere from 0 to N operations on that side of the queue,
rather than the "strictly 1" semantics of get() and put().